requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
attrs==25.3.0
cffi==1.17.1
discord.py==2.5.2
frozenlist==1.6.0
idna==3.10
multidict==6.4.3
pillow==11.2.1
propcache==0.3.1
pycparser==2.22
pygame==2.6.1
PyNaCl==1.5.0
yarl==1.20.0
